requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
django-celery-beat==2.4.0